Update tests
- Moved `valid_username` function into reddit.py
- Moved test for `valid_username` to test_reddit.py
- Added test for `lookup_table` in utils.py
- Removed dead code check in `reddit_from_env`. Tested locally.
