+++ tests/test_reddit.py
@@ -1,5 +1,18 @@
-from reddrec.reddit import Comments, hot_posts, usernames
+from reddrec.reddit import Comments, hot_posts, usernames, valid_username
 import numpy as np
+
+def test_valid_username():
+    # Proper symbols
+    assert valid_username('_example-123')
+    assert not valid_username('_*example-123')
+    assert not valid_username('spaces arent valid')
+
+    # Proper lengths
+    assert valid_username('lol')
+    assert valid_username('A' * 20)
+    assert not valid_username('xD')
+    assert not valid_username('A' * 21)
+    assert not valid_username('')
 
 def test_fetch_recent(i9n):
     with i9n.recorder.use_cassette('Comments.fetch_recent'):
